feat: inicio da rota de cadastro do usuario

+++ meu-primeiro-crud-go/src/controller/model/response/userResponse.go
@@ -4,5 +4,5 @@ type UserResponse struct {
 	Id    string `json:"id"`
 	Email string `json:"email"`
 	Name  string `json:"name"`
-	Age   int8   `json:"age"`
+	Age   int    `json:"age"`
 }
